app/backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.10.5
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-search-documents==11.4.0b11
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
Flask==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
msal==1.26.0
msal-extensions==1.1.0
packaging==23.2
portalocker==2.8.2
priority==2.0.0
pycparser==2.21
PyJWT==2.8.0
Quart==0.19.4
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
wsproto==1.2.0
